third_party/WebKit/LayoutTests/http/tests/security/powerfulFeatureRestrictions/durable-storage-on-insecure-origin.html

<!DOCTYPE html>
<title>Request Durable Storage On An Insecure Origin</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharness-helpers.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/get-host-info.js"></script>
<script>
if (window.location.origin != get_host_info().UNAUTHENTICATED_ORIGIN) {
    window.location = get_host_info().UNAUTHENTICATED_ORIGIN + window.location.pathname;
} else {
    promise_test(function() {
        return assert_promise_rejects(navigator.storage.persist(), 'SecurityError');
    }, "Requires secure context");
}
</script>
